Updates   * Got text componnt rendering styles properly   * Added JSDoc notes to some hooks   * Added proptypes
