+++ repos/keg-components/configs/aliases.json
@@ -15,7 +15,7 @@
   "KegSelect": "src/components/form/select/select.js",
   "KegSlider": "src/components/form/slider/slider${platform}.js",
   "KegSwitch": "src/components/form/switch/switch.js",
-  "KegText": "src/components/typography/kegtext${platform}.js",
+  "KegText": "src/components/typography/kegText/kegText.js",
   "KegView": "src/components/view/view${platform}.js",
   "KegGetPlatform": "src/utils/helpers/getPlatform.js",
   "KegAnimateHook": "src/hooks/useAnimate${platform}.js",
