+++ repos/keg-components/src/components/index.js
@@ -15,3 +15,4 @@ export * from './typography'
 export * from './view'
 export * from './modal'
 export * from './header'
+export * from './list'
