+++ repos/keg-components/src/hooks/index.js
@@ -6,6 +6,7 @@ export * from './usePressHandlers'
 export * from './useSelectHandlers'
 export * from './useSpin'
 export * from './useStyle'
+export * from './useTextAccessibility'
 export * from './useThemePath'
 export * from './useThemeWithHeight'
 export * from './useFromToAnimation'
