+++ repos/re-theme/src/styleInjector/styleInjector.js
@@ -1,4 +1,4 @@
-import React, { useMemo } from 'react'
+import React from 'react'
 import { useStyleTag } from './useStyleTag'
 
 /**
